# .scratch/mozhou-mvp/contracts/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "墨舟 MoZhou API",
    "version": "1.0.0",
    "description": "AI 小说写作平台 — 全站 19 组端点契约（UVSD Contract-Driven）"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
  "paths": {
...
    "/api/v1/skills": {
      "get": {
        "summary": "技能列表（mine/plaza）",
        "operationId": "skillList",
        "responses": {
          "200": {
            "description": "OK（scope=mine 返回 skills[]；scope=plaza 另返回 ohStorySkills[] 与 builtinSkills[]）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "skills": {
                      "type": "array"
                    },
                    "ohStorySkills": {
                      "type": "array"
                    },
                    "builtinSkills": {
                      "type": "array",
                      "description": "内置五角色技能定义（DELTA-004：toggleable 标记开关型——进胶囊由用户选择；基础设施型始终注入）",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": { "type": "string" },
                          "name": { "type": "string" },
                          "role": { "type": "string" },
                          "trigger": { "type": "string" },
                          "toggleable": { "type": "boolean" },
                          "connected": { "type": "boolean" },
                          "status": { "type": "string" },
                          "reason": { "type": "string", "nullable": true }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "mine",
                "plaza"
              ]
            }
          }
        ]
      },
      "post": {
        "summary": "创建/安装技能",
        "operationId": "skillCreate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "systemPrompt": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "description",
                  "systemPrompt"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除技能",
        "operationId": "skillDelete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/api/v1/sync/config": {
      "get": {
        "summary": "同步配置",
        "operationId": "syncGet",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "保存+连接测试",
        "operationId": "syncSave",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "autoSync": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "url",
                  "username",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/styles": {
      "get": {
        "summary": "获取我的风格库列表",
        "operationId": "listStyles",
        "responses": {
          "200": {
            "description": "当前用户全部风格（新→旧，无分页）"
          },
          "401": {
            "description": "未登录"
          }
        }
      },
      "post": {
        "summary": "保存风格指南",
        "operationId": "createStyle",
        "responses": {
          "201": {
            "description": "新建行摘要 {style:{id,name}}"
          },
          "400": {
            "description": "name 为空或 guide 非四维对象"
          },
          "401": {
            "description": "未登录"
          }
        }
      },
      "delete": {
        "summary": "删除我的风格",
        "operationId": "deleteStyle",
        "parameters": [
          {"name": "id", "in": "query", "required": true, "schema": { "type": "integer" }}
        ],
        "responses": {
          "200": {
            "description": "{ok:true}"
          },
          "400": {
            "description": "缺少风格 id"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "风格不存在或不归属当前用户"
          }
        }
      }
    },
...
  },
...
}